requirements.txt
acme==0.31.0
advertools==0.8.1
asn1crypto==0.22.0
blinker==1.3
cachetools==4.0.0
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2019.11.28
chardet==3.0.4
cloud-init==18.5
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
cryptography==1.9
decorator==4.4.2
fsspec==0.6.2
funcsigs==0.4
future==0.15.2
fuzzywuzzy==0.18.0
gcsfs==0.6.0
google-api-core==1.16.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.5
Jinja2==2.8
josepy==1.1.0
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
MarkupSafe==0.23
mock==1.3.0
ndg-httpsclient==0.4.2
nest-asyncio==1.3.0
numpy==1.18.1
oauthlib==3.1.0
pandas==0.24.2
pandas-gbq==0.13.1
parsedatetime==2.4
pbr==1.8.0
prettytable==0.7.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0
pydata-google-auth==0.3.0
pygobject==3.20.0
PyICU==1.9.2
PyJWT==1.3.0
PyMySQL==0.7.2
pyOpenSSL==17.3.0
pyparsing==2.0.3
pyRFC3339==1.0
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.8.1
python-debian==0.1.27
python-dotenv==0.12.0
python-systemd==231
pytz==2014.10
PyYAML==3.11
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
rsa==4.0
six==1.14.0
slack==0.0.2
slackclient==1.3.2
ssh-import-id==5.5
twython==3.7.0
ufw==0.35
urllib3==1.25.8
websocket-client==0.54.0
xlrd==1.2.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
